Move files into correct folders
