+++ app/components/WoWResults/ShortagePredictor/InputGenerator.tsx
